+++ discovery/src/main/java/org/openedx/discovery/data/model/ReviewDto.kt
@@ -8,7 +8,7 @@ data class ReviewDto(
     @SerializedName("profile_picture") val profilePicture: String?,
     @SerializedName("comment") val comment: String?,
     @SerializedName("designation") val designation: String?,
-    @SerializedName("rating") val rating: Int?,
+    @SerializedName("rating") val rating: Double?,
     @SerializedName("submitted_at") val submittedAt: String?,
 ) {
     fun mapToDomain() = Review(
@@ -16,7 +16,7 @@ data class ReviewDto(
         profilePicture = profilePicture.orEmpty(),
         comment = comment.orEmpty(),
         designation = designation.orEmpty(),
-        rating = rating ?: 0,
+        rating = (rating ?: 0.0).toDouble(),
         submittedAt = submittedAt.orEmpty()
     )
 }
